requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.0.post1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.34
Werkzeug==3.0.1
Jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3
